+++ containers/r/Singularity.r
@@ -3,7 +3,7 @@ Bootstrap: docker
 From: rocker/tidyverse
 
 %labels
-    version 0.2.0.2
+    version 0.2.0.3
 
 %environment
     OPENBLAS_NUM_THREADS=1
@@ -15,7 +15,8 @@ From: rocker/tidyverse
 
 %post
     apt-get update && apt-get install -y libhdf5-dev libssh-dev
+    R -e 'options(echo = TRUE); install.packages("minqa")' # weird problem where minqa can't install when .Rprofile isn't empty
     echo 'utils::setRepositories(ind=1:4); options(echo = TRUE)' > ~/.Rprofile
-    R --vanilla -e 'devtools::install_github("dynverse/dyndimred", dependencies = TRUE)'
-    R --vanilla -e 'devtools::install_github("dynverse/dynwrap@devel", dependencies = TRUE)'
-    R --vanilla -e 'devtools::install_cran(c("RcppEigen", "RSpectra", "RcppArmadillo"))' # preinstall certain rcpp libraries
+    R -e 'devtools::install_github("dynverse/dyndimred", dependencies = TRUE)'
+    R -e 'devtools::install_github("dynverse/dynwrap@devel", dependencies = TRUE)'
+    R -e 'devtools::install_cran(c("RcppEigen", "RSpectra", "RcppArmadillo"))' # preinstall certain rcpp libraries