+++ inst/NEWS.md
@@ -16,7 +16,9 @@
 
 * TESTING: Solve issue with the unit tests by not using any helpers.
 
-* MINOR CHANGE: Have docker images build from dynwrap@devel
+* MINOR CHANGE: Have docker images build from dynwrap@devel.
+
+* BUG FIX: Fix weird bug in Dockerfiles / recipes where minqa can't be installed when .Rprofile is not empty.
 
 # dynwrap 0.1.0 (07-03-2018)